Update plugin test
