+++ tests/plugins/test_platform_plugin.py
